probe: Implement PoC G29 command that has no effect on leveling yet
Added a G29 command that goes around the bed and probes a number of grid points.
For now, the variables are hard-coded, but it should be easy enough to add new
variables in the config. I still need to find a place to put the leveling logic,
which is up next. Another thing is that this code doesn't work if the absolute
Z-height isn't the lowest point on the bed, probing will fail when the height
falls below zero with a 'Timeout during homing' error.

Signed-off-by: Nick von Bulow <nick@nvonbulow.com>