probe: Implement basic z-height compensation and configuration for G29 probing
Created leveling algorithm classes for easy switching with the configuration
(much like the kinematics), but these classes are all contained within a
single file since these are simple enough. I may or may not have broken
delta and/or CoreXY kinematics with this commit since I only have a cartesian
printer, and I put the leveling object in the kinematics class since I need
access to absolute XY limits. Perhaps there's a better way to get the limits
fed to the leveling class, but this is what we have for now. Also, the first
position after homing is incorrect. For example, I created a test class that
adds one to all commanded z-heights, but the height after z probing is one
greater than it should be (0.3 mm when it should really be -0.7mm). This is
fixed at the first move after homing, so it shouldn't be critical, but it's
worth fixing.

Signed-off-by: nvonbulow <nick@nvonbulow.com>